// src/modules/auth/auth.schema.ts
import {email, z} from "zod"

export const registerSchema = z.object({
    name: z.string().min(2, "Name is too short"),
    email: z.string().email("Invalid email"),
    password: z.string().min(6, "Password must be atleast 6 characters"),
})

export const loginSchema = z.object({
    email: z.string().email("Invalid email"),
    password: z.string().min(6)
})

export type RegisterInput = z.infer<typeof registerSchema>
export type LoginInput = z.infer<typeof loginSchema>